Refactor Others folder tests into feature-based directories with proper isolation (#100)
* Initial plan

* refactor: reorganize Others folder tests into proper directories

- Move FileCommandsTests to FileOperations/ directory (3 files)
  - Base class + CreateEmpty + TestFile partials
  - All tests use unique files per test

- Move PowerQueryCommandsTests to PowerQuery/ directory (4 files)
  - Base class + Lifecycle + LoadConfig + Advanced partials
  - All tests use unique files via helper methods

- Move VbaTrustDetectionTests to VbaTrust/ directory (3 files)
  - Base class + TrustScope + ScriptCommands partials
  - Refactored to use CoreTestHelper for unique file creation

- Remove empty Others/ directory

All tests now follow established patterns:
- Each test uses unique Excel file for isolation
- Organized into feature-based directories
- Split into partial classes by feature area
- Validate both success/failure AND verify CRUD changes


---------
